+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file.
 
+## <a name="v0-6-1"></a> [0.6.1](https://github.com/bloodhunterd/backup/releases/tag/0.6.1) - 06.09.2020
+
+* Archive size and backup time added
+* Log and report messages fixed
+
 ## <a name="v0-6-0"></a> [0.6.0](https://github.com/bloodhunterd/backup/releases/tag/0.6.0) - 05.09.2020
 
 * Report layout overworked *(sections and emojis included)*